Store `redirect_uri` when creating authz code
RFC 6749 mandates that when the `redirect_uri` was provided in an
authorization request during the authorization code flow, it must
be provided again in the token request and the server needs to check
that the two match.

This change makes sure that we store the `redirect_uri` on the newly
created `AuthzCode` instance.